+++ src/case_runners/case_runner.jl
@@ -1,5 +1,5 @@
 function get_settings_path(case::AbstractString)
-    return joinpath(case, "Settings")
+    return joinpath(case, "settings")
 end
 
 function get_settings_path(case::AbstractString, filename::AbstractString)
@@ -7,7 +7,7 @@ function get_settings_path(case::AbstractString, filename::AbstractString)
 end
 
 function get_default_output_folder(case::AbstractString)
-    return joinpath(case, "Results")
+    return joinpath(case, "results")
 end
 
 @doc raw"""
@@ -51,10 +51,10 @@ function run_genx_case_simple!(case::AbstractString, mysetup::Dict, optimizer::A
     settings_path = get_settings_path(case)
 
     ### Cluster time series inputs if necessary and if specified by the user
-    TDRpath = joinpath(case, mysetup["TimeDomainReductionFolder"])
-
     if mysetup["TimeDomainReduction"] == 1
-        prevent_doubled_timedomainreduction(case)
+        TDRpath = joinpath(case, mysetup["TimeDomainReductionFolder"])
+        system_path = joinpath(case, mysetup["SystemFolder"])
+        prevent_doubled_timedomainreduction(system_path)
         if !time_domain_reduced_files_exist(TDRpath)
             println("Clustering Time Series Data (Grouped)...")
             cluster_inputs(case, settings_path, mysetup)
@@ -108,13 +108,14 @@ function run_genx_case_multistage!(case::AbstractString, mysetup::Dict, optimize
     mysetup["MultiStageSettingsDict"] = YAML.load(open(multistage_settings))
 
     ### Cluster time series inputs if necessary and if specified by the user
-    tdr_settings = get_settings_path(case, "time_domain_reduction_settings.yml") # Multi stage settings YAML file path
-    TDRSettingsDict = YAML.load(open(tdr_settings))
-
-    first_stage_path = joinpath(case, "Inputs", "Inputs_p1")
-    TDRpath = joinpath(first_stage_path, mysetup["TimeDomainReductionFolder"])
     if mysetup["TimeDomainReduction"] == 1
-        prevent_doubled_timedomainreduction(first_stage_path)
+        tdr_settings = get_settings_path(case, "time_domain_reduction_settings.yml") # Multi stage settings YAML file path
+        TDRSettingsDict = YAML.load(open(tdr_settings))
+    
+        first_stage_path = joinpath(case, "inputs", "inputs_p1")
+        TDRpath = joinpath(first_stage_path, mysetup["TimeDomainReductionFolder"])
+        system_path = joinpath(first_stage_path, mysetup["SystemFolder"])
+        prevent_doubled_timedomainreduction(system_path)
         if !time_domain_reduced_files_exist(TDRpath)
             if (mysetup["MultiStage"] == 1) && (TDRSettingsDict["MultiStageConcatenate"] == 0)
                 println("Clustering Time Series Data (Individually)...")
@@ -143,7 +144,7 @@ function run_genx_case_multistage!(case::AbstractString, mysetup::Dict, optimize
         mysetup["MultiStageSettingsDict"]["CurStage"] = t
 
         # Step 1) Load Inputs
-        inpath_sub = joinpath(case, "Inputs", string("Inputs_p",t))
+        inpath_sub = joinpath(case, "inputs", string("inputs_p",t))
 
         inputs_dict[t] = load_inputs(mysetup, inpath_sub)
         inputs_dict[t] = configure_multi_stage_inputs(inputs_dict[t],mysetup["MultiStageSettingsDict"],mysetup["NetworkExpansion"])
@@ -179,7 +180,7 @@ function run_genx_case_multistage!(case::AbstractString, mysetup::Dict, optimize
     end
 
     for p in 1:mysetup["MultiStageSettingsDict"]["NumStages"]
-        outpath_cur = joinpath(outpath, "Results_p$p")
+        outpath_cur = joinpath(outpath, "results_p$p")
         write_outputs(model_dict[p], outpath_cur, mysetup, inputs_dict[p])
     end