+++ src/multi_stage/write_multi_stage_network_expansion.jl
@@ -14,7 +14,7 @@ function write_multi_stage_network_expansion(outpath::String, settings_d::Dict)
     trans_capacities_d = Dict()
 
     for p in 1:num_stages
-        inpath = joinpath(outpath, "Results_p$p")
+        inpath = joinpath(outpath, "results_p$p")
         trans_capacities_d[p] = load_dataframe(joinpath(inpath, "network_expansion.csv"))
     end