+++ src/simple_operation.jl
@@ -1,4 +1,0 @@
-function simple_operation(num1::Float64, num2::Float64)
-    num3 = num1 + num2
-    return num3
-end