+++ test/test_electrolyzer.jl
@@ -5,7 +5,7 @@ using Test
 include(joinpath(@__DIR__, "utilities.jl"))
 
 obj_true = 6946.9618
-test_path = "Electrolyzer"
+test_path = "electrolyzer"
 
 # Define test inputs
 genx_setup = Dict(
@@ -24,9 +24,6 @@ end
 obj_test = objective_value(EP)
 optimal_tol_rel = get_attribute(EP, "ipm_optimality_tolerance")
 optimal_tol = optimal_tol_rel * obj_test  # Convert to absolute tolerance
-
-# Round the objective value to the same number of digits as the tolerance
-obj_test = round_objfromtol!(obj_test, optimal_tol)
 
 # Test the objective value
 test_result = @test obj_test ≈ obj_true atol = optimal_tol